+++ cli/cli/Commands/OtelCommands/OtelCommand.cs
@@ -4,7 +4,12 @@ namespace cli.OtelCommands;
 
 public class OtelCommand : CommandGroup
 {
-	public OtelCommand() : base("otel", "Allows access to Open Telemetry related commands")
+
+	public const string OTEL_COMMANDS_LOCK_FILE = "otel_operation";
+	
+	public OtelCommand() : base("telemetry", "Allows access to Open Telemetry related commands")
 	{
+		AddAlias("otel");
+		AddAlias("tel");
 	}
 }