+++ web/typedoc-plugin-markdown-fix.mjs
@@ -1,0 +1,8 @@
+import { MarkdownPageEvent } from 'typedoc-plugin-markdown';
+
+export function load(app) {
+  app.renderer.on(MarkdownPageEvent.END, (page) => {
+    // Remove backslashes before (e.g <T>) from page contents due to mkdocs rendering
+    page.contents = page.contents?.replace(/\\</g, '<');
+  });
+}