+++ cli/cli/CHANGELOG.md
@@ -5,6 +5,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Added `microServiceId` field to `MicroserviceServiceProviderRequest` to enable backend service grouping for proper microservice instance counting
+- `beam deploy plan` includes new `--docker-compose-directory` option to generate a docker compose project that can be used to run services locally
+- `beam project generate-env` no longer includes realm secret by default. Pass `--include-secret` to opt into realm secret. 
+- `beam project generate-env` can write a `.env` file of required environment variables by passing a `.env` file path to the command. 
+
+### Changed
+- Change `snapshots` for content now support multiple per realm snapshots for contents instead of only one main snapshot folder.
+- Event subscriptions use the developer's access token for authentication instead of the realm secret. 
+- `beam services` command suite is marked as internal
+- Rider IDE will be forced to re-run Beamable targets on project builds
+- `EnableUnrealBlueprintCompatibility` property is now added automatically to any `csproj` created in a Beamable project that is linked to an Unreal project.
+
+### Fixed
+- Fixed issue that would cause `content ps` if you had any non-`.json` files in any of your content folders.
+- Fixed issue in `beam unreal init` command that would cause it to fail if the provided path to the Unreal Sdk had a `.` in it.
+- Fixed issue with C#MS Static Analyzer that caused the `EnableUnrealBlueprintCompatibility` flag to be applied incorrectly.
+- Fixed issue with C#MS Static Analyzer when using the `EnableUnrealBlueprintCompatibility` flag that would incorrectly flag certain uses of generic types (`Promises`) as an error.  
+
 ## [5.4.2] - 2025-09-24
 ### Fixed
  - Issue when deserializing content with any fields that have the Attribute `[IgnoreContentField]`